[nuttx] Support nested signal handlers
This commit adds support for nested signal handlers to nuttx.
Its losely modeled on the same structure and nested syscalls.
Additionally this moves the pending signal queue from the task
group, into the task itself so that tasks cannot steal signals
from each other.

Finally it allocates tasks in vfork and create as the size of
a pthread_tcb.  This is because mono assumes the main thread
is a pthread.  This should probably be fixed by having the main
task pthread_create a target for mono, but this works around the
crash for now for a few bytes of memory.