Add support for M95/BG95 (#393)
Add support to Quectel M95 and Quectel BG95-M3 modules. Since each module has specific AT commands and turn-on procedures, a new field `Module` was added to cell.config.yaml, to allow Meadow to use different connect scripts and functions according to what was specified by the user.

The field Mode was modified to receive an string CATM1, NBIOT and GSM, instead of integers.

The turn-on functions are assuming that the GSM2 click and NimbeLink Skywire click boards are used, since a different click board may have different power-up requisites.