IPCP-provided DNS servers priority placement in dns.conf (#445)
**Context**
Some IP resolutions must be done by the IPCP-provided DNS servers, instead of the public ones. However, since there's a default DNS server, the 1.1.1.1, it was being used instead of the provided ones. To avoid making changes in an OS function, the approach used consists in reorder the DNS servers present in the DNS resolver file, letting the provided ones at the beginning of the dns.conf file, consequently, giving the highest priority during an IP resolution to them.

**Summary**
This PR aims to solve the following issue:
- https://github.com/WildernessLabs/Meadow/issues/444