Use Gpio to wakeup from low-power sleep Meadow_issue #271 (#446)
The STM32F7 has 6 wakeup pins designed for waking up the MCU from
low-power sleep and stop modes. However, none of these pins are
available on the FeatherV2 board. Therefore, the next best option is to
use an interrupt. Several types of interrupts can wakeup the MCU, but
this request is specifically about GPIO wakeup.
There are a few things to be aware of.
During low-power sleep, any interrupt will _partially_ wakeup the F7 and
call the assigned Interrupt Service Routine (ISR). However, this IRS
must start certain clocks etc. within the F7 to fully bring the F7 into
full operation. Therefore, if a GPIO not configured to wakeup the F7
generates a interrupt while the F7 is sleeping, the F7 is functionally
dead, only a restart will restore operation.
Therefore, the .Net app's creator must insure that, before calling
`Device.PlatformOS.Sleep` function, any GPIO that could create an
unwanted interrupt, must be disabled.
The only change to the .Net API is that an element in the struct used to
configure the GPIO was changed. The change was that the `enable` field
has been changed to `configType`. And the values previously accepted `0
= disable` or `1 = enable` are now `0 = remove`, `1 = add` and `2 =
wakeup` to declare that this GPIO is to be used to wakeup from low-power
sleep.