Adjust number of accepted ahdlc offline frames (#450)
It seems that increasing the number of accepted AHDLC offline frames,
which is related to the network timeout, caused some side effects, since
the OS takes much time to realize that it was disconnected (more than 15
minutes, sometimes), causing delays in the download. After doing large
OTA updates with poor cell networks, It appears that
reducing the timeout led to improved download performance.