Use single thread for CI build. (#456)
Building in GitHub is encountering the same build issue we encounter
sometime, namely missing object in the link phase. This change forces
the build in Github Actions CI to run as a single thread.

This is not intended as a permanent solution, simple a place holder
until the rebase project has completed and then this can be removed.