Removed App.exe test from mono startup, as it's tested by Meadow.dll (#468)
The test for `App.exe` has been in Meadow.OS for a very long time. I was
informed that CLI has been downloading both `App.dll` and CLI has been
renaming it to `App.exe` and downloading it too. I also was told that
Meadow.dll also tests for the correct `App.dll` file and won't start
Mono if it doesn't exist.