Corrected coding error in ESP32 download to address Meadow Issue #357 (#514)
The Meadow issue ESP32 Download was a simple coding error (see
Meadow_Issue #357 for details). It involved an enum in
`\apps\examples\hcom\file\hcom_file_dnld_esp32.c` contained 2 entries
both set to 1. After looking at the code I determined that this was
indeed an error. I made the change and retested the ESP32 code download.
Since it worked as designed I made the change to prevent confusion in
the future.