The recent addition of Subdirectory support in Meadow missed a few items. this PR address these items. (#519)
These are the main concerns that motivated this work:
1. If CLI user sends a single '/' character it is rejected by HCOM. I
should not be rejected but, be understood to be the root of the Meadow
fIle system.
2. There is no way to execute the `meadow file list` command and obtain
directory information.
3. Dominique found a ESP32 download bug. The message received is `No
active download, but data received`. After the message the download
fails. This has been fixed.

The cleanest way to resolved #1 and #2 above, and still support both
`CLIv1` and the other `CLIv2`, was to split the `file list` feature into
2 separate files, one supporting legacy `CLIv1` and the other supporting
`CLIv2`. The code for `CLIv1 ` has been isolated by a `#define
HCOM_FILE_LIST_SUPPORT_CLIV1_SCHEME` entry. This will simplify removing
this code when `CLIv1 `is retired. Two new commands were also added to
support the need for additional information for `CLIv2`. The HCOM
expected protocol version has been bumped to `8`. Protocol version `7`
is still accepted.