This PR covers a number of minor fixes and improvements for Subdirectory work, as well as SD-Card mounting/unmounting. (#526)
Because the SD-Card file system must be mounted for any access, this is
done automatically for `file write`, `file list`, `file delete` and
`file read`. As each SD-Card file system command is completed the
SD-Card is automatically unmounted.

This PR does NOT include changing the protocol version from 7 to 8. This
is because at present, users must use CLIv1 because it is the only
version available. And changing the protocol version would mean users
would get a message telling them to get the latest version of CLI, even
though it isn't available.

After this PR is merged, CLIv1 users should experience no difference in
behavior. They would be able to use `file write` and `file delete`, to
access the subdirectory features. However, since the CLIv1 user is not
able to access subdirectories using `file list` they would not be able
to examine the subdirectory structure, until CLIv2 is available.

A `#define HCOM_SUPPORT_CLIV1_LEGACY_BEHAVIOR` was added to make it easy
to remove unnecessary code with CLIv1 is retired.

All features are for both the `/meadow0` and `/sdcard` file systems.

To the best of my knowledge this completes the subdirectory additions.