Removed unused methods (#530)
Remove unused methods and clean up some of the comments.

### History:
The STM/ESP comms protocol uses a number of structures and associated
encoders and decoders to facilitate the transfer of data between the two
chips. There were generated automatically by a .NET application based
upon a meta data file. This was done while the system was undergoing
rapid change/development.

This is no longer the case and the files are now maintained manually.

### Purpose of the change
This change removes methods not used on the STM.

The protocol uses the following elements:

* Type containing the data
* Method to workout how much memory is needed to encode an instance of
the type
* Method to encode the data
* Method to extract the data

In all but a few cases, the STM needs to know how to encode a request
and extract the response. It does not need to know how to encode the
response or extract the request, these are needed by the ESP code but
not the STM code.