Rotary encoder upgrade for better and faster input handling, while reducing .Net load (#543)
Currently, interrupts generated by Nuttx are routed to `Meadow.Core` via
a memory queue (`mq`). This `mq `is not able to push interrupt
notifications to the managed code very quickly. This is very obvious
when using 1 or more rotary encoders. This is because the rotary encoder
generates a lot of interrupts for each rotation. At very slow speeds
everything is good, but as the speed increases notifications are lost
and the managed side looses information.